requirements.txt
app==0.0.1
apscheduler==3.11.2
fastapi==0.128.0
psycopg2_binary==2.9.10
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
